Unify cmdlets with parameter 'Encoding' to be of type System.Text.Encoding
Created support classes so tab completion would continue to wor
Create an ArgumentTransformationAttribute for the parameter to accept the currently used strings
Register encoding providers on all platforms. This seems safe enough, and there are defintely more
encodings available on CoreFX after this is done, which will be useful for some of the web cmdlet scenarios

Validate that files are created with UTF-8 encoding without BOM
Update tests to validate Encoding parameter to new type and create new tests for
parameter type validation.