Add right-to-left support to the -split operator
This commit adds the RightToLeft option to the Regex instance in SplitWithPattern() when the limit value specified by the user is negative. Using the RightToLeft options would ensure minimal overhead to existing cases with a positive or no limit.

SplitWithPredicate() is modified to inspect characters from right to left by introducing a cursor from which the strIndex derives its value. Works well but requires explicit checking and reversing of the string buffer before populating the resulting list when working right-to-left.