Change $OutputEncoding to use iso-8859-1 encoding rather than ASCII
This enables a number of useful scenarios.
With this change it is now possible to do the following:

```powershell
get-content -raw -encoding iso8859 /tmp/archive.tgz | gunzip | tar xvf -
```
and expand a compressed tar archive.
It also ensures that the bytes/characters passed to a native executable are not changed by the pipeline.
What is passed to the pipeline is passed to the next native application.