Added more infor as pe rthe review comment
One or more of the parameters issued cannot be used together.  This still doesn't address the case where too few parameters are supplied. For example:  foo  has 2 parameter sets  -A -B -C  and  -A -B -D  but the user only supplied  -A -B . It's not that  -A  and  -B  can't be used together, but that  -C  or  -D  are needed to determine the parameter set. - Added the information about not using the decision making parameter
• Do a Get-Help on your script/function/cmdlet for the syntax help.  I think this is would be somewhat confusing. I'm not sure how to fix it. The wording could at least be improved  Use Get-Help with your script, function, or cmdlet for syntax help.  - changed to suggested message