Documentation - Add to Mac OS install docs
Added a line documenting that the command to run PowerShell is `pwsh`, and not `powershell` like might be expected.
(I spent like 10 minutes trying to figure out why my fresh install wasn't working before I read further down in the manual install docs to find the correct command.)

Also, would it be worth it to add a command to the install process, making an `alias powershell='pwsh'`?  Or at least a line in the above docs?  That's what I ended up doing in my `.zshrc` for when future me forgets.