[Feature] Add functionality to update help in user scope
Update-Help needs elevation to update the help content.
This PR adds a functionality to save the updated help content under $HOME to elevation is not required.
To enable this functionality -Scope parameter is added Update-Help.
Scope is defaulted to Global, which adheres the current behavior, whereas User scope saves the help content under $HOME.
Also made changes to Get-Help to lookup help from under $HOME.
Updated and added tests for Update-Help and Get-Help.