[feature] Changed status code from string to int and using -Match instead of -Be to be more flexible
The WebListener requires the status code to be an int, therefore the test was changed to only use the integer form of the Status Code. As the Exception spells out the error message with spaces in the Status Code string, the -Be was changed to -Match and the String form of the Status Code removed.