Update PowerShell.Core.Instrumentation.man
Remove isolation attribute from Analytic and Debug channels. These channel types always have an isolation of Custom.

The Admin channel was enabled by default, but the comments for the channel indicated that it should be off by default and thus it is OK to use a huge file size and have it stop collecting events when the size is reached. This looks like a copy/paste error where the logging element was copied from a channel with a different isolation type, so fix it by duplicating the Operational channel's logging element instead.

Since no events are being written to the Admin channel currently, set it to disabled by default.