[Feature] Fix New-Item -Path with wildcard char
Unescape non-literal path when -Name is not set.
This works around for New-Item treating -Path as literal path while
it can also be globbable. For example, assuming there is "[file]1"
in current directory, and tab completion suggests './`[file`]1'
for the -Path argument, but
> New-Item -Path './`[file`]2'
will create a file named "`[file`]2" instead of "[file]2".