Fix type inference for foreach statement variable
Fix an issue where the assignment of a foreach statement would not be
detected during type inference. This issue was caused by a check
that the end offset of the foreach statement was before the start
offset of the variable expression. This was never true inside the
foreach statement body, which caused the assignement never to be
included.