[Feature] 2nd attempt to address review comments Major change is about how we decide to skip AMSI/malicious code scan.
A script block may comes from a .psd1 file but still has arbitrary code in it.
An attacker can parse an .psd1 file to get an Ast and call Ast.GetScriptBlock()
to create a script block. That script block comes from a .psd1 file that may
contain arbitrary content.

The skip-scan condition is changed to:
- the script block should come from a .psd1 file
- the script block is essentially a HashtableAst
- the HashtableAst is safe.