Remove unused cached types (#9015)
These types are not used in PowerShell code base and should be removed.

## PR Context

The dotnet 3.0.100-preview3-010406 version added overloads to `string.get_Chars` and hence we cannot resolved which overload to load. Thus, we get an type load exception.

Since we are no using this type, it is better to remove it.