Update the task-based async APIs added to PowerShell to return a Task object directly (#9079)
Update the task-based async APIs added to PowerShell to not use the `aysnc/await` keywords, but to return a `Task` object directly.
There is nothing to continue on after the `Task.Factory.FromAsync` call in those methods, so there is not need to use `aysnc` and `await` keywords, which turns the method into a state machine class unnecessarily.
# Conflicts:
#	src/System.Management.Automation/engine/hostifaces/PowerShell.cs
#	test/powershell/engine/Api/TaskBasedAsyncPowerShellAPI.Tests.ps1