Address feedback from @daxian-dbw
I don't need to keep the static strings of allowed module names or do convoluted checking of the experimental feature name.
Move the initialization of the HashSet with list of module names into the static constructor.
Ifdef a couple stopwatches out of the code (keeping the location tagged if we want to bring that back).
Add a comment about POWERSHELL_DISTRIBUTION_CHANNEL.
Change the type of the sessionid and uniqueid into strings so I don't repeatedly call ToString.