Tweak help formatting script to be more StrictMode tolerant
StrictMode is disabled in at least one other script block in the
help formatting. This change just adds another.

Addresses #10562

Turns out that just this one little addition of `Set-StrictMode -Off`
fixed the errors for my particular case.

(It is possible there are other ways to underspecify MAML, that would
require more, similar changes, but this looks like enough for me for
now.)