Address `--` and `--%` highlight issues
Correct tokenizing output so that `--` when used as a command name does not
instigate the 'stop processing parameters' function and is correctly
identified as a command.  Also correct further parameter tokens from being
incorrectly marked as commands while in that mode.
Correct tokenizing treatment of `--%` when it is a command name so as not
server as the VerbatimArgumentMode, and also to be indentified as a
command name token.
Address tokenizing of `--` (and others) as a function name, using
Token.SetIsCommandArgument.
This also addresses token flags being left set when operator tokens are
converted to generic tokens in Token.SetIsCommandArgument.
Also addressed assumptions that TokenKind.Generic is always StringToken.