Add MacPorts support to the build system
MacPorts is one of the "Missing package managers for OS X".
The original build scripts only supported Homebrew and it is inadvised to have
both installed on a same system.
This change makes the build script use whatever package manager is currently
installed, or direct the user to download one of them if neither is present.